requirements.txt
telethon==1.34.0
python-dotenv==1.0.0
requests==2.31.0
cryptg==0.4.0  # Optional: For faster encryption (recommended)